+++ package.json
@@ -16,10 +16,11 @@
     "@dnd-kit/core": "^6.1.0",
     "@dnd-kit/sortable": "^8.0.0",
     "@dnd-kit/utilities": "^3.2.2",
-    "@mantine/core": "^7.10.2",
+    "@mantine/core": "^7.12.1",
     "@mantine/dates": "^7.10.2",
     "@mantine/form": "^7.11.0",
     "@mantine/hooks": "^7.10.2",
+    "@mantine/notifications": "^7.12.1",
     "@neondatabase/serverless": "^0.9.1",
     "@prisma/adapter-neon": "^5.13.0",
     "@prisma/client": "^5.17.0",
